Changes requested during review
