+++ tests/components/schlage/snapshots/test_init.ambr
@@ -1,0 +1,33 @@
+# serializer version: 1
+# name: test_lock_device_registry
+  DeviceRegistryEntrySnapshot({
+    'area_id': None,
+    'config_entries': <ANY>,
+    'configuration_url': None,
+    'connections': set({
+    }),
+    'disabled_by': None,
+    'entry_type': None,
+    'hw_version': None,
+    'id': <ANY>,
+    'identifiers': set({
+      tuple(
+        'schlage',
+        'test',
+      ),
+    }),
+    'is_new': False,
+    'labels': set({
+    }),
+    'manufacturer': 'Schlage',
+    'model': '<model-name>',
+    'model_id': None,
+    'name': 'Vault Door',
+    'name_by_user': None,
+    'primary_config_entry': <ANY>,
+    'serial_number': None,
+    'suggested_area': None,
+    'sw_version': '1.0',
+    'via_device_id': None,
+  })
+# ---
